test(git): split the copier tests into worktree_copy_tree_test.go
The new regression pushed worktree_archive_test.go past the 1500-line
test limit. Split along the production seam instead of grandfathering it:
the copier's own tests now sit beside worktree_copy_tree.go, matching
worktree_copy_cleanup_test.go, and worktree_archive_test.go keeps the
move/publication flow. Pure code motion — no test bodies changed.
